[MCUX-74399] mimxrt700: openvg: enable the flash build targets on IAR
The OpenVG tiger example uses the XSPI1 PSRAM. The PSRAM is initialized
by the ROM when booting in flash mode only. For armgcc, the debug/release
targets already load into flash as well. For IAR, update the linker script
to use the flash and enable the flash_debug/flash_release build targets.
Also update the readme with instructions on how to set the boot mode on the
board.

Signed-off-by: Camelia Groza <camelia.groza@nxp.com>