[MCUX-73237] add note to fix host cannot be resumed issue.
When usb pd shield boardis installed, the connected pins are working in
1.8 power domain. However, the shield board gives 3.3V to the connected
GPIO pin of evkmimxrt685. This will cause evkmimxrt685 works abnormally
then cause the resume issue. In this case, need to change the related
jumper setting to set the right power domain from 1.8V to 3.3V for the
connected GPIOs. This note is also applied for mimxrt685audevk.

Signed-off-by: Aiden Hu <weiwei.hu@nxp.com>