[WSW-55014][RT1060][ncp_examples] Increase the size of the console input buffer
[Description]
When the data input in the host console is too long at one time, such as: 'ot dataset set active [leader dataset data]'. It will occurs 'Error: Ring buffer overrun...'.

[Root Cause]
The reason for this error is that the UART receiving buffer is too small, and the data is not processed in time, resulting in an overflow of the receiving buffer.

[Fix]
Increase the size of the UART receiving buffer.

[Testing]
Pass

Signed-off-by: Allen Dong <allen.dong@nxp.com>