[MCUX-73518] Fixed "task creation failed" in lwip_ipv4_ipv6_echo example.
The shell was used in blocking mode and it did not let run the IDLE task,
which caused the memory of self-deleted tasks not to be freed.

Fixed by using non-blocking shell and adjusting heap memory size.

As a consequence of using non-blocking shell and debug console,
some prints started to be out of order, e.g. shell prompt being printed
before the end of the shell command's output.
To have the correct order of output, some PRINTF calls have been
replaced with SHELL_Printf where possible, and added debug console
flush where replacing by SHELL_Printf was not possible.

Signed-off-by: Stanislav Poboril <stanislav.poboril@nxp.com>