[MCUX-76431][2/5] Improve performance of example dev_cdc_ecm_nic.
This patch only improve the Ethernet frame procession logic.

Improvements:
- Use systick to tigger the link status check for minimizing redundant
calls.
- Add USB (OUT direction) receive request into receive response
callback, so that the device always sends data to the host as
uninterrupted as possible.
- Use variable instead of function access ENET's MDIO and MDC to
minimize instruction execution times.
- Discard continuous receive or continuous transmit both of USB and
ENET, so that the demo as a transmission medium is as unobstructed as
possible in both directions.
- Using the SPSC model & lock-free ring buffer to free up transmission
performance as much as possible.

Signed-off-by: Dv Alan <zhangyang.shen@nxp.com>