[WSW-54031][RW612][NCP] Fix board reset issue after iperf is done in PM3
[Description]
After PM3 is enabled on NCP device, start iperf test. The board will
reset when iperf is done.

[Root Cause]
The reset is caused by ITRC of gdet sensor. The itrc interrupt was not
disabled due to one extra calling of function
POWER_DisableGDetVSensors(). The ncp_intf_pm_enter returned fail when
iperf is done and the uart is still busy. There is no status check in
kPM_EventExitingSleep case and POWER_DisableGDetVSensors was called for
the second time but this time, this API did nothing because
s_gdetSensorContext.disableCount is not 0. This parameter only becomes
zero after actual resuming from PM3.

[Fix]
Added flag board_notify_done. It will be set to true when the board
enter case is fully done. In exit case, check this flag. If exiting from
PM3 and the flag is true, do the board reinit. Before exiting, reset the
flag to false.

[Testing]
No board reset was found after iperf is done in PM3.

[Reviwers]
Gannicus Guo, Gang Li

Signed-off-by: Hui Bai <hui.bai@nxp.com>