[MCUX-77982] change cmd select_conn as select_ag
The cmd is used to select the ag connection for
the other cmds, it is not used to select the acl conn
because there may be many acl conns, but EdgeFast only
support 2 ag connections, so it is used to select the
ag connection.
change the select index as 0 and 1.

Signed-off-by: Mark Wang <yichang.wang@nxp.com>