[MCUX-78279][rt700] vglite_examples: Update the power management code
- use 325MHz media domain clock, PMIC power supply, for better performance

- disable LDO:
When running at the max 325MHz frequency, VDD2 needs to be 1.1V.
The internal LDO for VDD2 can only go up to 1.05V per the datasheet,
so an external PMIC is required to supply the 1.1V for VDD2.
The vglite_examples  turn on the PMIC and set the output to 1.1V,
but they don't then switch to PMIC mode and turn off the internal LDOs.
See Section 27.6.2.1.3 in the RT700 RM.

- specify about power supply settings in the application readmes
The project still (usually) works when JP1 and JP3 are open if LDO is
not disabled. It can sometimes be a bit flakey (since it's out of spec)
but it will still usually run correctly. If the power scheme was set up
correctly though, then it shouldn't work at all with JP1 and JP3 OPEN
since there would be no power for VDD2.

Signed-off-by: Andrei Catalin Ilie <andreicatalin.ilie@nxp.com>