[MCSE-735] Rework and remove unnecessary ADC APIs
ADC is deinitialized before low power entry and it requires to be reinit after low power.
The reason why the ADC is deinitialized on low power entry is to reduce the power consumption (saves 15uA).
After The ADC reinit, the module should perform the calibrations needed, which takes more than 3.5ms. Meaning
that if we want to trigg the temperature measurement on wakeup we should keep the host awake for at least 3.5ms
in order to recalibrate the ADC + adding time to trigg and receive the measurement from ADC by ISR. This is unacceptable.

To skip this constraint, we propose to save the calibration data, deinit ADC before low power entry, and restore this
calibration after wakeup. This doesn't work with current code since `LPADC_SetCalibrationValue` waits for calibration
mode to be ready `ADC_STAT_CAL_RDY_MASK`. This bit gets the value 1 after ADC init + calibration but it loses
its state during the ADC deinit, to be more precise after disabling the ADC clock. And this is what causes
`LPADC_SetCalibrationValue` to hang after low power exit.

The workaround is to perform offset calibration after low power exit, which sets the `ADC_STAT_CAL_RDY_MASK` to 1 and
will allow us then to use `LPADC_SetCalibrationValue` to set the saved calibration data (saved before low power entry)

changes apply to:
-mcxw72loc