MCUX-80195 Handle errata 011063 for MCXE247 power_mode_switch.
Errata: When the bus clock is same system clock and an asynchronous wakeup occurs during a mode transition from RUN to VLPS or VLPR to VLPS, the MCU may hang in an undetermined state, which can only be recovered by a power-on reset event or a watchdog reset.
Applied workaroud: Before executing the transition to VLPS ensure that the PREDIV_SYS_CLK frequency / BUS_CLK frequency configuration for RUN/VLPR mode is greater than or equal to 2.

Signed-off-by: Holt.Sun <holt.sun@nxp.com>