[WSW-60083][MATTER-NCP] Fix issue about send mDNS fail
[Description]
Not find correct netif when do mDNS process

[Root Cause]
When do mDNS the ipv6 DAD process not complete, so can't get the correct netif to mDNS, which cause mDNS send pkg fail.

[Fix]
Add an event to notify the ipv6_state turn to prefered so that ncp-host and MATTER side can get the information about DAD complete.

[Testing]
Test passed with MATTER-NCP

Signed-off-by: Xinyu Hong <xinyu.hong@nxp.com>