[MCSE-864][wireless_nbu] Fix wrong threadx initializaion in low-level assembly files
The handler priorities were not correctly set due to r0 pointing to
the _tx_thread_system_stack_ptr symbol instead of the SCB base address.
This caused several scheduling issues.
Fixed it by re-ordering the instructions, now the setup of the stack
pointer is done first, and all the SCB configuration is done after, to
avoid having to change r0 multiple times.

Signed-off-by: Axel Le Bourhis <axel.lebourhis@nxp.com>