[WSW-61512][rw612][ncp tpt] usb throughput optimization pass buff as a pointer and remove memcpy to improve thoughput results
[Description]
usb throughput optimization pass buff as a pointer and remove memcpy to improve thoughput results
[Root Cause]
At ncp interface layer the api ncp_usb_device_recv copy the recv data to a global arrays by memcpy.
To improve the thoughtput result, remove the memcpy process, and directly use the pointer of recv_data to pass it to the next level API, that is, continue to use the buff pointed to by recv_data without re-malloc buff to save this data.
Based on this change, the throughput of the USB interface can be increased from 15Mbps to 25Mbps.
[Fix]
Rewrite the implementation of ncp_usb_device_recv api, use the recv_data pointer to pass data directly, and no longer re-malloc buff to save this data
[Testing]
Pass with ncp examples

Signed-off-by: Xinyu Hong <xinyu.hong@nxp.com>