[HCIOT-5081][MCXW23] Fix LL initiating Tx before the UART FIFO is empty
The LL hci_transport layer has a state machine which sequentially sends
parts of the hci packets and waits for a callback from hci_uart before
sending the next piece of information.

hci_uart was only relying on DMA Tx interrupt to mark the buffer as
transmitted and notify hci_transport. This is not correct since DMA only
transfers the buffer to the UART FIFO but the data still needs to be sent
through the UART line.
The LL was therefore notified to soon and initiated a new transfer which
resulted in bytes from the previous transfer never being sent.
This was especially noticeable at baud rates < 230400 bps.

The patch modifies hci_uart to wait for the all the bytes to be sent
before notifying hci_transport.