[MCUX-82567][MCXW23] Fix apps compilation on MacOS
Some MCXW23 BLE related apps like hci_bb_mcxw23 use a #define to initialize a static array
which can be board dependent.
This is conveniently done in a reconfig.cmake file which is specific to the app and the board.

This define is rather complex (array of structs) such as
BLECONTROLLER_GPIO_MAP={\
    {DBG_IO_SCHDLR_HNDL_NXT_TRACE,10},\
    {DBG_IO_LLWCC_CMN_HG_ISR,13},\
    {DBG_IO_SCHDLR_EXEC_EVNT_TRACE,14},\
    {DBG_IO_SCHDLR_HNDL_MISSED_EVNT,15}\
}
and is not accepted by every preprocessor, giving the following error
<command-line>: error: invalid initializer

Using quotes around the outer braces seems to solve the issue but when using them in the reconfig.cmake
the system strips them out (also tried with escaping)

This define is not needed for the functional code but just for debugging so it can be safely removed
to make compilation pass.