[MCUX-82420] lwip_iperf_bm: decreased default UDP rate to 50 Mbit/s.
Decreased IPERF_UDP_CLIENT_RATE. When the rate is too high, board may
be overwhelmed by the iperf traffic and miss the ending datagram.
This can happen in situations when the other side (PC) is sending UDP
to a board at the client rate determined by a board (UDP client
reverse, dual and tradeoff modes).

Signed-off-by: Stanislav Poboril <stanislav.poboril@nxp.com>