[MCSE-1127][health_care_iot_peripheral] Increase the watchdog timeout to 3s for BM build
PSA enablement increases the time to generate keys during BLE pairing.
During this time, even though the health task has a higher priority than
the BLE Host task, it can't preempt the running task to feed the watchdog.
This can cause the watchdog to trigger, so we increase the timeout to 3 seconds.
We keep 2s in FreeRTOS builds.

Signed-off-by: Axel Le Bourhis <axel.lebourhis@nxp.com>