[WSW-62599]Modify the input of the crc calculation process to four-byte alignment
[Description]
Fix crc check error of ncp.
[Root Cause]
At ncp-host side, during dcp calculation process, the dcp logic pads the zero to a 32-bit boundary for the trailing bytes, and the zero will participate in calculations.
At ncp-device, the crc process will not add zero to input to a 32-bit boundary, so that the calculate result will different with ncp-host side.
The reason why the verification of wlan-version can work properly is that it is four-byte aligned and does not need to be padded with 0 through DCP.
[Fix]
Modify the input of the crc calculation process to four-byte alignment by padded with 0 at ncp deivce side.
[Testing]
Pass with ncp examples

Signed-off-by: Xinyu Hong <xinyu.hong@nxp.com>