[MCUX-83351] Fix dev_suspend_resume_hid_mouse no output issue
This issue only occurs on FRDM-K32L3A6, and the root cause is that this
case configure GPIOA's PCR before related clock enabled that made the
GPIO hard fault.

There is the risk of button GPIO is not configurationed at pin mux,
and the button GPIO is configured confusedly at hardware_init.c.
Since the feature of button change before, that made the initialization
flow of GPIO is incorrect.

Correct the pin mux configruations and discard the redundant button GPIO
settings in hardware_init.c.

Signed-off-by: Dv Alan <zhangyang.shen@nxp.com>