[MCUX-83547] Fix example dev_dfu_freertos download issue on RD-RW612-BGA
This issue occurs on debug target at ARMGCC or MDK toolchain. The root
cause is that the code footprint of the dev_dfu_freertos exceeds the
preset address 0x10000 (reserved for firmware). When this example
attempts to download firmware, the firmware overwrites
dev_dfu_freertos's code space, causing a hard fault and resulting in
DFU download failure.

Increased the preset firmware address offset and updated the firmware
to fix this issue.

Signed-off-by: Dv Alan <zhangyang.shen@nxp.com>