[WSW-67374][Redfinch][NCP][UART][Supplicant]Ping fail to extAP while fast ping from AP backend to DUT
[Description]
Ping fail to extAP while fast ping from AP backend to DUT
[Root Cause]
On the ncp-host side, after sending a ping request, the socket task continuously receives ping requests from the backend while waiting to receive a ping reply. These requests are discarded after check it if they are not the required ping reply.
However, there is a timeout mechanism in this process. If 10 consecutive packets are received that are not ping replies, the task will not wait anymore and sends the next ping request.
In this case, due to the backend's continuous fast pinging, more than 10 ping requests from the backend will be received within the waiting time after the ncp-host sends a ping request. This causes the ncp-host's socket task to be unable to properly receive and parse ping replies.
[Fix]
At ncp-host side, remove the retry count.
[Testing]
Test pass with ncp examples
[Reviewers]
Gannicus Guo

Signed-off-by: Xinyu Hong <xinyu.hong@nxp.com>